Add files via upload
This Health monitoring system allows users to keep a track of their body vitals like oxygen level and heartbeat then remotely send
the data to healthcare professionals. The data which is collected by the sensor is displayed on the LED screen and at the same time this data is fetched by the Blynk IoT platform. We can remotely access the data from anywhere through this.